# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponseDto" }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        },
        "summary": "Login with email and password",
        "tags": ["auth"]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponseDto" }
              }
            }
          },
          "400": { "description": "Bad request" }
        },
        "summary": "Register a new user with organization",
        "tags": ["auth"]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponseDto" }
              }
            }
          },
          "401": { "description": "Invalid refresh token" }
        },
        "summary": "Refresh access token using refresh token",
        "tags": ["auth"]
      }
    },
    "/api/auth/check-token": {
      "post": {
        "operationId": "AuthController_checkToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Token status information" },
          "401": { "description": "Invalid refresh token" }
        },
        "summary": "Check refresh token status (diagnostics)",
        "tags": ["auth"]
      }
    },
...
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": { "204": { "description": "Logged out successfully" } },
        "security": [{ "bearer": [] }],
        "summary": "Logout (invalidate current session)",
        "tags": ["auth"]
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "204": { "description": "Logged out from all devices successfully" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Logout from all devices (invalidate all sessions)",
        "tags": ["auth"]
      }
    },
    "/api/auth/session": {
      "get": {
        "operationId": "AuthController_getSession",
        "parameters": [],
        "responses": {
          "200": { "description": "User session info retrieved successfully" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get current user session info",
        "tags": ["auth"]
      }
    },
    "/api/auth/user-settings": {
      "post": {
        "operationId": "AuthController_updateUserSettings",
        "parameters": [],
        "responses": {
          "200": { "description": "Settings updated successfully" },
          "400": { "description": "Bad request" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update user and organization settings",
        "tags": ["auth"]
      }
    },
...
    "/api/plugins/proxy/{pluginId}/{path}": {
      "get": {
        "operationId": "PluginProxyController_proxyRequest_get",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      },
      "post": {
        "operationId": "PluginProxyController_proxyRequest_post",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      },
      "put": {
        "operationId": "PluginProxyController_proxyRequest_put",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      },
      "delete": {
        "operationId": "PluginProxyController_proxyRequest_delete",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      },
      "patch": {
        "operationId": "PluginProxyController_proxyRequest_patch",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      },
      "options": {
        "operationId": "PluginProxyController_proxyRequest_options",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      },
      "head": {
        "operationId": "PluginProxyController_proxyRequest_head",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      },
      "search": {
        "operationId": "PluginProxyController_proxyRequest_search",
        "parameters": [
          {
            "name": "pluginId",
            "required": true,
            "in": "path",
            "description": "Plugin ID",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "description": "Path to proxy",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Proxy request to plugin server",
        "tags": ["plugin-proxy"]
      }
    },
...
  },
  "info": {
    "title": "eTickets Platform API",
    "description": "The eTickets Platform API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          }
        },
        "required": ["email", "password"]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string" },
          "permissions": { "type": "array", "items": { "type": "string" } },
          "organizationId": { "type": "string" },
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn": { "type": "number" }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "permissions",
          "organizationId",
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "User full name"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "organizationName": {
            "type": "string",
            "example": "My Organization",
            "description": "Organization name"
          },
          "organizationSlug": {
            "type": "string",
            "example": "my-organization",
            "description": "Organization slug (auto-generated if not provided)"
          },
          "role": {
            "type": "string",
            "example": "owner",
            "description": "User role within the organization"
          }
        },
        "required": ["name", "email", "password", "organizationName"]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn": { "type": "number" }
        },
        "required": ["accessToken", "refreshToken", "expiresIn"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token received from login or registration"
          }
        },
        "required": ["refreshToken"]
      },
...
    }
  }
}